% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sbgjackets_dependencies.R
\name{check_sbgjackets_dependencies}
\alias{check_sbgjackets_dependencies}
\title{Check  \code{sbgjackets} run-time system dependencies}
\usage{
check_sbgjackets_dependencies()
}
\description{
\code{check_sbgjackets_dependencies()} checks this \code{sbgjackets}
run-time system dependencies.
}
\details{
\code{sbgjackets} has the following run-time system dependencies:
\enumerate{
\item The Carlito font.
\item R compiled with support for the Cairo-devices i.e. \code{cairo_pdf()}.
\item System tools to embed various pdf metadata such as \code{exiftool}, \code{gs}, and/or \code{pdftk}.
}
}
