% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_grobs.R
\name{spineTextGrob}
\alias{spineTextGrob}
\alias{creditsGrob}
\alias{backNotesGrob}
\alias{spineIconGrob}
\title{Small Box Game Jacket Grobs}
\usage{
spineTextGrob(title, col = "white", size = c("4x6", "poker"))

creditsGrob(
  xmp = xmpdf::xmp(),
  credits = character(),
  icons = FALSE,
  size = "4x6",
  ...,
  width = NULL,
  x = NULL
)

backNotesGrob(
  notes,
  col = "black",
  size = c("4x6", "poker"),
  ...,
  style = sbgjackets_style(size, color = col)
)

spineIconGrob(players, minutes, weight, col = "white", size = "4x6")
}
\arguments{
\item{title}{Title of small box game jacket (usually game name)}

\item{col}{Color of text/icons}

\item{size}{Target box size.  Either \code{"4x6"} or \code{"poker"}.}

\item{xmp}{A \code{\link[xmpdf:xmp]{xmpdf::xmp()}} object with copyright/license information.}

\item{credits}{A character vector of (commonmark) credits to eventually be
passed to \code{\link[marquee:marquee_grob]{marquee::marquee_grob()}}.}

\item{icons}{If \code{TRUE} include Creative Commons credits for Games-icons.net icons.}

\item{...}{Ignored for now.}

\item{x, width}{\code{\link[grid:unit]{grid::unit()}} that will be passed to \code{\link[marquee:marquee_grob]{marquee::marquee_grob()}}.}

\item{notes}{A character vector of (commonmark) back cover notes.}

\item{style}{A \code{marquee} style set object
(e.g. from \code{\link[=sbgjackets_style]{sbgjackets_style()}}).}

\item{players}{A integer vector of allowed number of players}

\item{minutes}{An integer of number of minutes (larger number from BGG)}

\item{weight}{A double of game weight (number from BGG)}
}
\value{
A grid grob object
}
\description{
\code{backNotesGrob()}, \code{creditsGrob()}, \code{spineIconGrob()}, and
\code{spineTextGrob()} are helper grob functions to help make small
box game jackets using \code{\link[pnpmisc:pdf_create_jacket]{pnpmisc::pdf_create_jacket()}}.
}
\examples{
library("grid", include.only = c("gList", "gpar"))
sbgj_example <- function(output = NULL) {
  front <- fullGrob("#D55E00")
  back <- fullGrob("#009E73")
  spine <- gList(fullGrob("black"),
                 spineTextGrob("Example Spine"),
                 spineIconGrob(2:4, 30, 1.5))
  inner <- creditsGrob(icons = TRUE)
  pnpmisc::pdf_create_jacket(output, front = front, back = back,
                             spine = spine, inner = inner)
}
}
