% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/everdeck.R
\name{pcbj_everdeck}
\alias{pcbj_everdeck}
\title{Create playing card box jacket for the Everdeck}
\usage{
pcbj_everdeck(
  output = NULL,
  ...,
  paper = getOption("papersize", "letter"),
  instructions = FALSE
)
}
\arguments{
\item{output}{Output file name.  Defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Should be empty.}

\item{paper}{Paper size.  Usually either "letter" or "a4" but in certain circumstances can be one of "letter", "a4", "poker", "bridge", "legal", "a5", "a3", "special".}

\item{instructions}{If \code{TRUE} then prepend instructions on how to make the jacket to the beginning of the pdf}
}
\value{
The output file name invisibly.  As a side effect creates a pdf file.
}
\description{
\code{pcbj_everdeck()} creates two playing card box jackets for the Everdeck.
}
