% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{pcbj_cubes}
\alias{pcbj_cubes}
\alias{pcbj_dice}
\alias{pcbj_polyhedral_dice}
\title{Create playing card box jackets for generic component storage.}
\usage{
pcbj_cubes(
  output = NULL,
  ...,
  paper = getOption("papersize", "letter"),
  instructions = FALSE
)

pcbj_dice(
  output = NULL,
  ...,
  paper = getOption("papersize", "letter"),
  instructions = FALSE
)

pcbj_polyhedral_dice(
  output = NULL,
  ...,
  paper = getOption("papersize", "letter"),
  instructions = FALSE
)
}
\arguments{
\item{output}{Output file name.  Defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Should be empty.}

\item{paper}{Paper size.  Usually either "letter" or "a4" but in certain circumstances can be one of "letter", "a4", "poker", "bridge", "legal", "a5", "a3", "special".}

\item{instructions}{If \code{TRUE} then prepend instructions on how to make the jacket to the beginning of the pdf}
}
\value{
The output file name invisibly.  As a side effect creates a pdf file.
}
\description{
\code{pcbj_cubes()} creates a playing card box jacket for wooden cube storage.
\code{pcbj_dice()} creates a playing card box jacket for dice storage.
\code{pcbj_polyhedral_dice()} creates a playing card box jacket for polyhedral dice storage.
}
