% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs.r
\name{pcbj_wizard}
\alias{pcbj_wizard}
\title{Create playing card box jacket for \emph{Wizard}}
\usage{
pcbj_wizard(
  output = NULL,
  ...,
  paper = getOption("papersize", "letter"),
  instructions = FALSE,
  double = FALSE
)
}
\arguments{
\item{output}{Output file name.  Defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Should be empty.}

\item{paper}{Paper size.  Usually either "letter" or "a4" but in certain circumstances can be one of "letter", "a4", "poker", "bridge", "legal", "a5", "a3", "special".}

\item{instructions}{If \code{TRUE} then prepend instructions on how to make the jacket to the beginning of the pdf}

\item{double}{If \code{TRUE} produce two jackets instead of one jacket.}
}
\value{
The output file name invisibly.  As a side effect creates a pdf file.
}
\description{
\code{pcbj_wizard()} creates a playing card box jacket for the game \emph{Wizard}.
}
\details{
Note that these print-and-play playing card box jackets are for \strong{Personal Use Only}.
}
