% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blorg_in_the_midwest.R
\name{pcw_blorg_in_the_midwest}
\alias{pcw_blorg_in_the_midwest}
\title{Create card wallet jacket for \emph{Blorg in the Midwest}}
\usage{
pcw_blorg_in_the_midwest(
  output = NULL,
  ...,
  paper = getOption("papersize", "letter")
)
}
\arguments{
\item{output}{Output file name.  Defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Should be empty.}

\item{paper}{Paper size.  Usually either "letter" or "a4" but in certain circumstances can be one of "letter", "a4", "poker", "bridge", "legal", "a5", "a3", "special".}
}
\value{
The output file name invisibly.  As a side effect creates a pdf file.
}
\description{
\code{pcw_blorg_in_the_midwest()} creates a card wallet jacket for the game
\emph{Blorg in the Midwest}.
}
