% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_polish_jacket.R
\name{pdf_polish_jacket}
\alias{pdf_polish_jacket}
\title{Polish a jacket pdf with instructions and/or metadata}
\usage{
pdf_polish_jacket(output, ..., instructions = FALSE, xmp = NULL)
}
\arguments{
\item{output}{Filename of a pdf file created by \code{\link[pnpmisc:pdf_create_jacket]{pnpmisc::pdf_create_jacket()}} or \code{\link[pnpmisc:pdf_create_jacket]{pnpmisc::pdf_create_poker_jacket()}}.}

\item{...}{Should be empty.}

\item{instructions}{If \code{TRUE} then prepend instructions on how to make the jacket to the beginning of the pdf}

\item{xmp}{A \code{\link[xmpdf:xmp]{xmpdf::xmp()}} object.}
}
\value{
The \code{output} filename invisibly
}
\description{
\code{pdf_polish_jacket()} polishes a jacket pdf by adding instructions and/or metadata.
}
