% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gipf.r
\name{sbgj_yinsh}
\alias{sbgj_yinsh}
\title{Create SBG Jacket for YINSH}
\usage{
sbgj_yinsh(
  output = NULL,
  ...,
  paper = getOption("papersize", "letter"),
  instructions = FALSE
)
}
\arguments{
\item{output}{Output file name.  Defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Should be empty.}

\item{paper}{Paper size.  Usually either "letter" or "a4" but in certain circumstances can be one of "letter", "a4", "poker", "bridge", "legal", "a5", "a3", "special".}

\item{instructions}{If \code{TRUE} then prepend instructions on how to make the jacket to the beginning of the pdf}
}
\value{
The output file name invisibly.  As a side effect creates a pdf file.
}
\description{
\code{sbgj_yinsh()} creates a small box game jacket for \href{https://www.gipf.com/yinsh/}{YINSH}.
}
\details{
Note that these print-and-play small box game jackets are for \strong{Personal Use Only}.
}
