% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.r
\name{sbgj_storage_all}
\alias{sbgj_storage_all}
\alias{sbgj_dice}
\alias{sbgj_glass_stones}
\alias{sbgj_pawns}
\alias{sbgj_polyhedral_dice}
\title{Create SBG Jackets for Generic Component Storage}
\usage{
sbgj_storage_all(
  output = NULL,
  ...,
  paper = c("letter", "a4"),
  instructions = TRUE
)

sbgj_dice(output = NULL, ..., paper = c("letter", "a4"), instructions = FALSE)

sbgj_glass_stones(
  output = NULL,
  ...,
  paper = c("letter", "a4"),
  instructions = FALSE
)

sbgj_pawns(output = NULL, ..., paper = c("letter", "a4"), instructions = FALSE)

sbgj_polyhedral_dice(
  output = NULL,
  ...,
  paper = c("letter", "a4"),
  instructions = FALSE
)
}
\arguments{
\item{output}{Output file name.  Defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Should be empty.}

\item{paper}{Paper size.  Either "letter", "a4", or "special".}

\item{instructions}{If \code{TRUE} then prepend instructions on how to make the jacket to the beginning of the pdf}
}
\value{
The output file name invisibly.  As a side effect creates a pdf file.
}
\description{
\code{sbgj_glass_stones()} creates a small box game jacket for glass stones.
\code{sbgj_pawns()} creates a small box game jacket for pawns.
\code{sbgj_polyhedral_dice()} creates a small box game jacket for polyhedral dice.
\code{sbgj_storage_all()} creates all of those into a single pdf file.
}
