% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbgjackets_style.R
\name{sbgjackets_style}
\alias{sbgjackets_style}
\title{Small Box Game Jackets \code{marquee} style}
\usage{
sbgjackets_style(
  size = c("4x6", "poker", "wallet"),
  ...,
  color = "black",
  align = NULL,
  cex = 1
)
}
\arguments{
\item{size}{Either \code{"4x6"} for a 4x6 photo storage box jacket,
\code{"poker"} for a playing card storage box, or
\code{"wallet"} for a card wallet origami jacket.}

\item{...}{Ignored}

\item{color}{The desired text color.}

\item{align}{If not \code{NULL} the text alignment
(e.g. \code{"left"}, \code{"center"}, or \code{"right"})
to use for the \code{"base"} style,
see \code{\link[marquee:style]{marquee::style()}} for allowed values.}

\item{cex}{Multiplier applied to the base font size (and hence the
line spacing which is relative to the font size).}
}
\value{
A \code{marquee} style set object.
}
\description{
\code{sbgjackets_style()} returns a \code{marquee} style set object
to help add text to jackets in a consistent style.
}
