% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traditional_card_games.r
\name{pcbj_bridge}
\alias{pcbj_bridge}
\alias{pcbj_pinochle}
\title{Create playing card box jacket for traditional card games}
\usage{
pcbj_bridge(
  output = NULL,
  ...,
  paper = c("letter", "a4"),
  instructions = FALSE
)

pcbj_pinochle(
  output = NULL,
  ...,
  paper = c("letter", "a4"),
  instructions = FALSE
)
}
\arguments{
\item{output}{Output file name.  Defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Should be empty.}

\item{paper}{Paper size.  Either "letter", "a4", or "special".}

\item{instructions}{If \code{TRUE} then prepend instructions on how to make the jacket to the beginning of the pdf}
}
\value{
The output file name invisibly.  As a side effect creates a pdf file.
}
\description{
\code{pcbj_bridge()} creates a playing card box jacket for a Bridge deck.
\code{pcbj_pinochle()} creates a playing card box jacket for a Pinochle deck.
}
