% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{spineTextGrob}
\alias{spineTextGrob}
\alias{creditsGrob}
\alias{spineIconGrob}
\title{Helper grobs}
\usage{
spineTextGrob(title, col = "white")

creditsGrob(xmp = xmpdf::xmp(), credits = character(), icons = FALSE)

spineIconGrob(players, minutes, weight, col = "white")
}
\arguments{
\item{title}{Title of small box game jacket (usually game name)}

\item{col}{Color of text/icons}

\item{xmp}{A \code{\link[xmpdf:xmp]{xmpdf::xmp()}} object with copyright/license information.}

\item{credits}{A character vector of (commonmark) credits to eventually be
passed to \code{\link[marquee:marquee_grob]{marquee::marquee_grob()}}.}

\item{icons}{If \code{TRUE} include Creative Commons credits for Games-icons.net icons.}

\item{players}{A integer vector of allowed number of players}

\item{minutes}{An integer of number of minutes (larger number from BGG)}

\item{weight}{A double of game weight (number from BGG)}
}
\value{
A grid grob object
}
\description{
\code{creditsGrob()}, \code{spineIconGrob()}, and \code{spineTextGrob()}
are helper grob functions to help make small box game jackets
using \code{\link[pnpmisc:pdf_create_jacket]{pnpmisc::pdf_create_jacket()}}.
}
\examples{
library("grid", include.only = c("gList", "gpar", "rectGrob"))
sbgj_example <- function(output = NULL) {
  front <- rectGrob(gp = gpar(col = NA, fill = "#D55E00"))
  back <- rectGrob(gp = gpar(col = NA, fill = "#009E73"))
  spine <- gList(rectGrob(gp = gpar(col = NA, fill = "black")),
                 spineTextGrob("Example Spine"),
                 spineIconGrob(2:4, 30, 1.5))
  inner <- creditsGrob(icons = TRUE)
  pnpmisc::pdf_create_jacket(output, front = front, back = back,
                             spine = spine, inner = inner)
}
}
