% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/everdeck.r
\name{pcbj_everdeck}
\alias{pcbj_everdeck}
\title{Create playing card box jacket for the Everdeck}
\usage{
pcbj_everdeck(
  output = NULL,
  ...,
  paper = c("letter", "a4"),
  instructions = FALSE
)
}
\arguments{
\item{output}{Output file name.  Defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Should be empty.}

\item{paper}{Paper size.  Either "letter", "a4", or "special".}

\item{instructions}{If \code{TRUE} then prepend instructions on how to make the jacket to the beginning of the pdf}
}
\value{
The output file name invisibly.  As a side effect creates a pdf file.
}
\description{
\code{pcbj_everdeck()} creates two playing card box jackets for the Everdeck.
}
