% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foxtrot_games.r
\name{pcbj_fox_in_the_forest}
\alias{pcbj_fox_in_the_forest}
\title{Create playing card box jacket for \emph{Fox in the Forest}}
\usage{
pcbj_fox_in_the_forest(
  output = NULL,
  ...,
  paper = c("letter", "a4"),
  instructions = FALSE
)
}
\arguments{
\item{output}{Output file name.  Defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Should be empty.}

\item{paper}{Paper size.  Either "letter", "a4", or "special".}

\item{instructions}{If \code{TRUE} then prepend instructions on how to make the jacket to the beginning of the pdf}
}
\value{
The output file name invisibly.  As a side effect creates a pdf file.
}
\description{
\code{pcbj_fox_in_the_forest()} creates a playing card box jacket for the game \emph{Fox in the Forest}.
}
\details{
Note that these print-and-play playing card box jackets are for \strong{Personal Use Only}.
}
