% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playing_cards.R
\name{pcbj_bavarian_pattern}
\alias{pcbj_bavarian_pattern}
\alias{pcbj_castilian_pattern}
\alias{pcbj_english_pattern}
\title{Create playing card box jacket for various playing card decks.}
\usage{
pcbj_bavarian_pattern(
  output = NULL,
  ...,
  paper = getOption("papersize", "letter"),
  instructions = FALSE,
  double = FALSE
)

pcbj_castilian_pattern(
  output = NULL,
  ...,
  paper = getOption("papersize", "letter"),
  instructions = FALSE,
  double = FALSE
)

pcbj_english_pattern(
  output = NULL,
  ...,
  paper = getOption("papersize", "letter"),
  instructions = FALSE,
  double = FALSE
)
}
\arguments{
\item{output}{Output file name.  Defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Should be empty.}

\item{paper}{Paper size.  Usually either "letter" or "a4" but in certain circumstances can be one of "letter", "a4", "poker", "bridge", "legal", "a5", "a3", "special".}

\item{instructions}{If \code{TRUE} then prepend instructions on how to make the jacket to the beginning of the pdf}

\item{double}{If \code{TRUE} produce two jackets instead of one jacket.}
}
\value{
The output file name invisibly.  As a side effect creates a pdf file.
}
\description{
\code{pcbj_bavarian_pattern()} creates a playing card box jacket for a German-suited deck in the Bavarian pattern.
\code{pcbj_castilian_pattern()} creates a playing card box jacket for a Spanish-suited deck in the Castilian pattern.
\code{pcbj_english_pattern()} creates a playing card box jacket for a French-suited deck in the English pattern.
}
