% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominoes.r
\name{sbgj_dominoes_all}
\alias{sbgj_dominoes_all}
\alias{sbgj_dominoes_double6}
\alias{sbgj_dominoes_double9}
\alias{sbgj_dominoes_double12}
\title{Create SBG Jacket for Dominoes}
\usage{
sbgj_dominoes_all(output = NULL, ..., paper = c("letter", "a4"))

sbgj_dominoes_double6(output = NULL, ..., paper = c("letter", "a4"))

sbgj_dominoes_double9(output = NULL, ..., paper = c("letter", "a4"))

sbgj_dominoes_double12(output = NULL, ..., paper = c("letter", "a4"))
}
\arguments{
\item{output}{Output file name.  Defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Should be empty.}

\item{paper}{Paper size.  Either "letter", "a4", or "special".}
}
\value{
The output file name invisibly.  As a side effect creates a pdf file.
}
\description{
\code{sbgj_dominoes_double6()}, \code{sbgj_dominoes_double9()}, and \code{sbgj_dominoes_double12()} creates a small box game jacket for Double-Six Dominoes, Double-Nine Dominoes, and Double-Twelve Dominoes.
\code{sbgj_dominoes_all()} creates all of those into a single pdf file.
}
