% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.r
\name{sbgj_dice}
\alias{sbgj_dice}
\alias{sbgj_marbles}
\alias{sbgj_glass_stones}
\alias{sbgj_pawns}
\alias{sbgj_polyhedral_dice}
\title{Create SBG Jackets for Generic Component Storage}
\usage{
sbgj_dice(
  output = NULL,
  ...,
  paper = getOption("papersize", "letter"),
  instructions = FALSE
)

sbgj_marbles(
  output = NULL,
  ...,
  paper = getOption("papersize", "letter"),
  instructions = FALSE
)

sbgj_glass_stones(
  output = NULL,
  ...,
  paper = getOption("papersize", "letter"),
  instructions = FALSE
)

sbgj_pawns(
  output = NULL,
  ...,
  paper = getOption("papersize", "letter"),
  instructions = FALSE
)

sbgj_polyhedral_dice(
  output = NULL,
  ...,
  paper = getOption("papersize", "letter"),
  instructions = FALSE
)
}
\arguments{
\item{output}{Output file name.  Defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Should be empty.}

\item{paper}{Paper size.  Usually either "letter" or "a4" but in certain circumstances can be one of "letter", "a4", "poker", "bridge", "legal", "a5", "a3", "special".}

\item{instructions}{If \code{TRUE} then prepend instructions on how to make the jacket to the beginning of the pdf}
}
\value{
The output file name invisibly.  As a side effect creates a pdf file.
}
\description{
\code{sbgj_glass_stones()} creates a small box game jacket for glass stones.
\code{sbgj_marbles()} creates a small box game jacket for marbles.
\code{sbgj_pawns()} creates a small box game jacket for pawns.
\code{sbgj_polyhedral_dice()} creates a small box game jacket for polyhedral dice.
}
