% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbgjackets_style.r
\name{sbgjackets_style}
\alias{sbgjackets_style}
\title{Small Box Game Jackets \code{marquee} style}
\usage{
sbgjackets_style(size = c("4x6", "poker"), ..., color = "black")
}
\arguments{
\item{size}{Either \code{"4x6"} for a 4x6 photo storage box jacket or \code{"poker"} for a playing card storage box.}

\item{...}{Ignored}

\item{color}{The desired text color.}
}
\value{
A \code{marquee} style set object.
}
\description{
\code{sbgjackets_style()} returns a \code{marquee} style set object
to help add text to jackets in a consistent style.
}
