% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traditional_card_games.r
\name{pcbj_bridge}
\alias{pcbj_bridge}
\alias{pcbj_pinochle}
\alias{pcbj_poker}
\title{Create playing card box jacket for traditional card games}
\usage{
pcbj_bridge(
  output = NULL,
  ...,
  paper = getOption("papersize", "letter"),
  instructions = FALSE,
  double = FALSE
)

pcbj_pinochle(
  output = NULL,
  ...,
  paper = getOption("papersize", "letter"),
  instructions = FALSE,
  double = FALSE
)

pcbj_poker(
  output = NULL,
  ...,
  paper = getOption("papersize", "letter"),
  instructions = FALSE,
  double = FALSE
)
}
\arguments{
\item{output}{Output file name.  Defaults to \code{tempfile(fileext = ".pdf")}.}

\item{...}{Should be empty.}

\item{paper}{Paper size.  Usually either "letter" or "a4" but in certain circumstances can be one of "letter", "a4", "poker", "bridge", "legal", "a5", "a3", "special".}

\item{instructions}{If \code{TRUE} then prepend instructions on how to make the jacket to the beginning of the pdf}

\item{double}{If \code{TRUE} produce two jackets instead of one jacket.}
}
\value{
The output file name invisibly.  As a side effect creates a pdf file.
}
\description{
\code{pcbj_bridge()} creates a playing card box jacket for a Bridge deck.
\code{pcbj_pinochle()} creates a playing card box jacket for a Pinochle deck.
\code{pcbj_poker()} creates a playing card box jacket for a Poker deck.
}
